Fix: project timezone
Extension needs to be aware of project timezone in order to set the correct sun dates. Note that we don't need to go the extra step via UTC since both the date returned from the Forma API and the dates created within the extension have the same local timezone. We just need to know what the difference between the project timezone and that is.

Depends on:
* https://github.com/spacemakerai/forma-extensions/pull/127
* https://github.com/spacemakerai/forma-embedded-view-sdk/pull/51
* https://github.com/spacemakerai/designmode/pull/1253

Have verified that it works for projects in Europe, California and New York with the above PRs wired up locally.