Create Date objects with Luxon (#3)
* Create Date objects with Luxon

The implementation in https://github.com/autodesk-platform-services/aps-forma-extension-shadow-study/pull/1 had at least one flaw:

It calculated the offset based on the `currentDate` from the Forma API.

But there is no guarantee that the offset is the same for the user selected date, as it is for the currently selected shadow date in Forma.

This is particularly true for Daylight Savings Time, observed in many parts of the world.

But why would DST change the offset when looking at native javascript Date, who by default is in the clients locale?
Because different terretories observe DST on different dates.

An example of this:
* Norway stops observing DST Sunday 29th of October 2023
* East coast of the US stops observing DST at Sunday 5th of November 2023.

This means that with me as a client, the "getTimezoneOffset" is different for that particular week.
There is a similar period occuring during the spring.

The actual effect of this is that for certain combinations of user locales, and date selection, the selected hour range would be off by one. The files creates still wrote it out with `timeZone`, so any written screenshots would be correctly named.

I opted to use a Luxon instead of wrangling with things like these. There are ways to work with offsets like the initial implementation tried, but in general I feel it is very easy to forget various edge cases. I also prefer instantiating the date in a given timezone.
If someone has strong preferences to use something else than Luxon, and I'm all ears.

I guess we could also have `getTimezoneOffset` work by properly passing the selected date. But then again the "selected" date might depend on the timezone offset.

Fun fact about time zones:
Did you know that there is a 15 minute time difference between India and Nepal? Nepal is UTC + 5:45 :O

* Update readme to reflect changes