Avoid windows reserverd characters (#8)
The approach with "toLocale" may include `:` or `/` in file names inside the zip file. Windows fails to unzip these files with errors around "The parameter is incorrect".

Instead just format to simle HH-mm regardless of locale.

I've verified this works as it should across timezones, where I still got the appropriate file names on a project in the US, while my machine had CET as timezone.